Move retrieving element attributes to utilities
