+++ lib/plugins/event-tracker.js
@@ -17,11 +17,9 @@
 
 var assign = require('object-assign');
 var delegate = require('dom-utils/lib/delegate');
-var getAttributes = require('dom-utils/lib/get-attributes');
 var provide = require('../provide');
-var camelCase = require('../utilities').camelCase;
 var createFieldsObj = require('../utilities').createFieldsObj;
-
+var getAttributeFields = require('../utilities').getAttributeFields;
 
 /**
  * Registers declarative event tracking.
@@ -69,21 +67,7 @@ EventTracker.prototype.handleEvents = function(event, element) {
   // Ensures the event type matches the one specified on the element.
   if (event.type != element.getAttribute(prefix + 'on')) return;
 
-  var defaultFields = {};
-  var attributes = getAttributes(element);
-
-  Object.keys(attributes).forEach(function(attribute) {
-    if (attribute.indexOf(prefix) === 0 && attribute != prefix + 'on') {
-      var value = attributes[attribute];
-
-      // Detects Boolean value strings.
-      if (value == 'true') value = true;
-      if (value == 'false') value = false;
-
-      var field = camelCase(attribute.slice(prefix.length));
-      defaultFields[field] = value;
-    }
-  });
+  var defaultFields = getAttributeFields(element, prefix);
 
   this.tracker.send(defaultFields.hitType || 'event', createFieldsObj(
       defaultFields, this.opts.fieldsObj, this.tracker, this.opts.hitFilter));
