+++ lib/utilities.js
@@ -16,6 +16,7 @@
 
 
 var assign = require('object-assign');
+var getAttributes = require('dom-utils/lib/get-attributes');
 
 
 var utilities = {
@@ -69,6 +70,38 @@ var utilities = {
 
 
   /**
+   * Retrieves the attributes from an DOM element and returns a fields object
+   * for all attributes matching the passed prefix string.
+   * @param {Element} element The DOM element to get attributes from.
+   * @param {string} prefix An attribute prefix. Only the attributes matching
+   *     the prefix will be returned on the fields object.
+   * @return {Object} An object of analytics.js fields and values
+   */
+  getAttributeFields: function(element, prefix) {
+    var attributes = getAttributes(element);
+    var attributeFields = {};
+
+    Object.keys(attributes).forEach(function(attribute) {
+
+      // The `on` prefix is used for event handling but isn't a field.
+      if (attribute.indexOf(prefix) === 0 && attribute != prefix + 'on') {
+
+        var value = attributes[attribute];
+
+        // Detects Boolean value strings.
+        if (value == 'true') value = true;
+        if (value == 'false') value = false;
+
+        var field = utilities.camelCase(attribute.slice(prefix.length));
+        attributeFields[field] = value;
+      }
+    });
+
+    return attributeFields;
+  },
+
+
+  /**
    * Accepts a function and returns a wrapped version of the function that is
    * expected to be called elsewhere in the system. If it's not called
    * elsewhere after the timeout period, it's called regardless. The wrapper
@@ -90,6 +123,12 @@ var utilities = {
   },
 
 
+  /**
+   * Accepts a string containing hyphen or underscore word separators and
+   * converts it to camelCase.
+   * @param {string} str The string to camelCase.
+   * @return {string} The camelCased version of the string.
+   */
   camelCase: function(str) {
     return str.replace(/[\-\_]+(\w?)/g, function(match, p1) {
       return p1.toUpperCase();
